update payload and forwarder.  Add 1.1.0 release. Fix extraction for the pack with overwrite ini.

+++ amssu-forwarder/source/main.cpp
@@ -15,27 +15,28 @@ int main()
 	std::cout<< "CREATE DIRECTORY PATH" << std::endl;
 	consoleUpdate(NULL);
 	std::filesystem::create_directory(PATH);
-	if(std::filesystem::exists(BAD_PATH)){
+	if(std::filesystem::exists(BAD_PATH) && std::filesystem::exists(FULL_PATH)){
 		std::cout<< "REMOVE BAD PATH" << std::endl;
 		consoleUpdate(NULL);
 		std::filesystem::remove(BAD_PATH);
 	}
 
-
+	if(std::filesystem::exists(BAD_PATH) && !std::filesystem::exists(FULL_PATH)){
+		std::cout<< "REMOVE BAD PATH" << std::endl;
+		consoleUpdate(NULL);
+		std::filesystem::rename(BAD_PATH, FULL_PATH);
+	}
 
 	if(std::filesystem::exists(APP_DOWNLOAD_PATH)) {
 		std::cout<< "RENAME DOWNLOAD PATH" << std::endl;
 		consoleUpdate(NULL);
 		std::filesystem::remove(FULL_PATH);
 		std::filesystem::rename(APP_DOWNLOAD_PATH, FULL_PATH);
+		std::filesystem::remove(APP_DOWNLOAD_PATH);
 	}
 
 	std::cout << "REMOVES" << std::endl;
 	consoleUpdate(NULL);
-		if(std::filesystem::exists(APP_DOWNLOAD_PATH)) {
-
-		std::filesystem::remove(APP_DOWNLOAD_PATH);
-	}
 
 	std::filesystem::remove(FORWARDER_PATH);
